Update README.md
During the MODAL473V, we got the chance to implement a U-Net to detect ships for the Airbus Ship detection kaggle challenge. 

Do not hesitate to read the "rendu final" which goes more into detail about how and why we built our model, as well as performance issues and fixes.